Gassama

+++ src/main/java/org/psychohelp/psychohelp/service/ResultatService.java
@@ -5,7 +5,7 @@ import org.psychohelp.psychohelp.entity.ResultatTest;
 import java.util.List;
 
 public interface ResultatService {
-    ResultatTest calculerEtEnregistrerResultat(Integer citoyenId, Integer testId, List<Integer> choixIds);
+    ResultatTest calculerEtEnregistrerResultat(Long citoyenId, Integer testId, List<Integer> choixIds);
 
-    List<ResultatTest> obtenirResultatsParCitoyen(Integer citoyenId);
+    List<ResultatTest> obtenirResultatsParCitoyen(Long citoyenId);
 }
